+++ src/main/java/janggi/Application.java
@@ -1,6 +1,12 @@
 package janggi;
 
 import janggi.domain.*;
+import janggi.domain.game.GameDao;
+import janggi.domain.game.GameService;
+import janggi.domain.piece.PieceDao;
+import janggi.domain.piece.PieceService;
+import janggi.domain.turn.TurnDao;
+import janggi.domain.turn.TurnService;
 
 public class Application {
